Fix contact form captcha verification
when 'Login attempts' settings was set to non-zero value, then captcha
verification on contact form was always skipped.

The logic for contact form should not be mixed with login-flow logic,
which can include several attempts without captcha.

Contact form should be always secured with captcha, unless customer is
logged in.

Closes #5